fix: rename pkg.components.link to pkg.navigation.link; fix build error
