+++ apps/xi.front/package.json
@@ -33,7 +33,6 @@
     "next-sitemap": "^3.1.42",
     "notistack": "^3.0.0-alpha.11",
     "npm": "^9.2.0",
-    "pkg.components.link": "*",
     "pkg.components.loading": "*",
     "pkg.components.scroll": "*",
     "pkg.data.avatar": "*",
@@ -73,6 +72,7 @@
     "pkg.icons.updates": "*",
     "pkg.icons.users": "*",
     "pkg.layout.pages": "*",
+    "pkg.navigation.link": "*",
     "pkg.pages.error": "*",
     "pkg.theme.scheme": "*",
     "prop-types": "^15.8.1",
